refactor: aggregate errors by types
