+++ ciphering-tool/errors/base-error.js
@@ -1,6 +1,17 @@
+/**
+ * @param { Error } error
+ * @return { string }
+ */
+const getMessage = (error) => error.message.replace(/.*:\s/, '');
+
 export class BaseError extends Error {
-  constructor(message) {
-    super(message);
+  /**
+   * @param { Error | string } error
+   * @param { string } [prefix]
+   */
+  constructor(error, prefix) {
+    const message = error instanceof Error ? getMessage(error) : error;
+    super(`${prefix}: ${message}`);
     this.name = this.constructor.name;
   }
 }
