--- ciphering-tool/errors/cli-errors/cli-error-no-arg.js
@@ -1,7 +1,0 @@
-import { CliError } from './cli-error.js';
-
-export class CliErrorNoArg extends CliError {
-  constructor() {
-    super('no arguments');
-  }
-}
