--- ciphering-tool/errors/cli-errors/cli-error-option-duplicate.js
@@ -1,11 +1,0 @@
-import { CliErrorArg } from './cli-error-arg.js';
-
-export class CliErrorOptionDuplicate extends CliErrorArg {
-  /**
-   * @param {number} position
-   * @param {string} option
-   * */
-  constructor(position, option) {
-    super(position, option, 'duplicated option');
-  }
-}
