--- ciphering-tool/errors/cli-errors/cli-error.js
@@ -1,8 +1,0 @@
-import { BaseError } from '../base-error.js';
-
-export class CliError extends BaseError {
-  /** @param {string} message */
-  constructor(message) {
-    super(`Command Error: ${message}`);
-  }
-}
