--- ciphering-tool/errors/cli-errors/index.js
@@ -1,4 +1,0 @@
-export { CliErrorNoArg } from './cli-error-no-arg.js';
-export { CliErrorUnknownArg } from './cli-error-unknown-arg.js';
-export { CliErrorOptionWithoutArg } from './cli-error-option-without-arg.js';
-export { CliErrorOptionDuplicate } from './cli-error-option-duplicate.js';
