--- ciphering-tool/errors/config-errors/config-error-invalid-token.js
@@ -1,7 +1,0 @@
-import { ConfigError } from './config-error.js';
-
-export class ConfigErrorInvalidToken extends ConfigError {
-  constructor(position, code) {
-    super(`on position "${position + 1}" - invalid code "${code}"`);
-  }
-}
