--- ciphering-tool/errors/config-errors/config-error-no-config.js
@@ -1,7 +1,0 @@
-import { ConfigError } from './config-error.js';
-
-export class ConfigErrorNoConfig extends ConfigError {
-  constructor() {
-    super('no config');
-  }
-}
