--- ciphering-tool/errors/config-errors/config-error.js
@@ -1,8 +1,0 @@
-import { BaseError } from '../base-error.js';
-
-export class ConfigError extends BaseError {
-  /** @param {string} message */
-  constructor(message) {
-    super(`Config Error: ${message}`);
-  }
-}
