--- ciphering-tool/errors/config-errors/index.js
@@ -1,2 +1,0 @@
-export { ConfigErrorNoConfig } from './config-error-no-config.js';
-export { ConfigErrorInvalidToken } from './config-error-invalid-token.js';
