--- ciphering-tool/errors/file-error/file-error-no-name.js
@@ -1,7 +1,0 @@
-import { FileError } from './file-error.js';
-
-export class FileErrorNoName extends FileError {
-  constructor() {
-    super('no file name');
-  }
-}
