--- ciphering-tool/errors/file-error/file-error-not-readable.js
@@ -1,7 +1,0 @@
-import { FileError } from './file-error.js';
-
-export class FileErrorNotReadable extends FileError {
-  constructor(name) {
-    super(`not readable file "${name}"`);
-  }
-}
