--- ciphering-tool/errors/file-error/file-error-not-writable.js
@@ -1,7 +1,0 @@
-import { FileError } from './file-error.js';
-
-export class FileErrorNotWritable extends FileError {
-  constructor(name) {
-    super(`not writable file "${name}"`);
-  }
-}
