--- ciphering-tool/errors/file-error/file-error.js
@@ -1,8 +1,0 @@
-import { BaseError } from '../base-error.js';
-
-export class FileError extends BaseError {
-  /** @param {string} message */
-  constructor(message) {
-    super(`File System Error: ${message}`);
-  }
-}
