--- ciphering-tool/errors/file-error/index.js
@@ -1,3 +1,0 @@
-export { FileErrorNoName } from './file-error-no-name.js';
-export { FileErrorNotReadable } from './file-error-not-readable.js';
-export { FileErrorNotWritable } from './file-error-not-writable.js';
