+++ ciphering-tool/errors/stream-error.js
@@ -1,0 +1,8 @@
+import { BaseError } from './base-error.js';
+
+export class StreamError extends BaseError {
+  /** @param { Error | string } error */
+  constructor(error) {
+    super(error, 'Stream Error');
+  }
+}
